Remove defaults for browser version
With the 'version' parameter specified and running in hub mode, the hub
is unable to find a matching client and rejects the request.

According to https://github.com/SeleniumHQ/selenium/wiki/DesiredCapabilities
the default should be an empty string if not known.
Additionally the 'browserVersion' is not a valid capability.