Bump to current Behat 3.6.1
Also update other dependencies to actual ones.

Special mention to symfony/process raised to 4.x or 5.x
so it only can run with php71 or newer.

Hence, this only can be backported to Moodle 3.7 and up.

We also need to ensure that we stay using behat/mink < 1.8
in this branch because 1.8 came with php74 support but also
with some session incompatible changes that only are
available for 38_STABLE and up.