Failed Steps should always cause a failure
Some types of failed step were ignored and not detected as having either
an Exception, or a Failure result, leading to a false pass result for
that step.

To give a specific example, FailedStepSearchResult is returned when a
Step cannot be found, or when multiple steps match the same search
string.

The FailedStepSearchResult is a StepResult, which does have an exception
and whose `isPassed` method returns false (to indicate a failure),
however the class is not an ExecutedStepResult, and it not an
UndefinedStepResult, or a SkippedStepResult. As a result it is believed
to not support Chained Steps at all, and the result is entirely replace
with a new Exception Test step thereby causing the step to be marked,
incorrectly, as Passed.

This patch adjusts the checking to explicitly check for the presence of:
- any exception; or
- any test failure.

The StepTester should always return a StepResult, which is a type of
TestResult.

Any TestResult which _can_ have any exception _must_ implement the
ExceptionResult interface, and will return a true for the `hasException`
call.

Any TestResult which _can_ have a result _must_ implement the StepResult
Interface, and _must_ return a boolean value for the `isPassed` call.

Using this we can therefore check whether any Step which can either
throw an exception, or be marked as failed, is already in a failure
state and stop testing immediately.

Other TestResult types then fit into two categories:
- those which do not contain any Chained Step; and
- those which make use of the deprecated Chained Step.

For those TestResult instances which _do not_ support chaining, a new step is
performed to check for Exceptions, and only the result of that one test
is returned.

For those TestResult instances which _do_ support chaining, all child
steps are executed.