Switch (temporarily) to own fork of friends-of-behat/mink-extension
To get some PHP8 warnings removed, by applying:

https://github.com/FriendsOfBehat/MinkExtension/pull/9

Once that PR has been accepted upstream then
this commit can be reverted and we can be back to use the upstream
repository friends-of-behat/mink-extension, or, more yet, if all this
is applied to the original behat/mink-extension we could also move back
to it.

We'll remove our own fork once any of the upstream ones are
working ok.