Set the default value for Marionette to True
Selenium 4+ validates Capabilities now, and will not accept a Null value
for a Boolean property (marionette).

The real issue here is that the Selenium2Factory is very opionated and
defaults to Firefox with a load of properties.

This is the only one which is really problematic as the rest have either
no default, or sensible defaults.

Where a property has no default, it is considered optional, and is
removed. Where a default is specified, then it will be applied.

There is no way to entirely remove a property once specified using the
Builder. The only way is to specify a new sensible default.